+++ src/js/components/AddCoursePanel/AddCoursePanel.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import { contextConsumer } from '../ApiContext/ContextConsumer';
+import { contextConsumer } from '../../context/apiContext';
 import SearchPanel from '../SearchPanel';
 import Button from '../Button';
 import styles from './AddCoursePanel.less';
