--- src/js/components/ApiContext/ApiContext.js
@@ -1,3 +1,0 @@
-import React from 'react';
-
-export const ApiContext = React.createContext();
