--- src/js/components/ApiContext/ContextConsumer.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import { ApiContext } from './ApiContext';
-
-export const contextConsumer = Component => props => (
-  <ApiContext.Consumer>
-    {context => <Component {...props} context={context} />}
-  </ApiContext.Consumer>
-);
