--- src/js/components/ApiContext/index.js
@@ -1,3 +1,0 @@
-import ApiContext from './ApiContext';
-
-export default ApiContext;
